improve handle clock
